Updated base model for verification. Create utils_base for difference in utility functions.

+++ data_loader.py
@@ -44,4 +44,4 @@ def import_data(folder_path, data_name):
 
         print('To-do: Need to check for standardisation!')
 
-        return X, y
+        return X.astype('float32'), y.astype('float32')
